Help form Conner
Conner and I were working on algorithms and had extra time after, looked at Mason and I's code for practicum to see if we could improve using information from two groups. By comparing to Conner's code, and working together, we made progress on my code, so the client is closer to being finished and tested to make sure that server works. Conner, Max, y'all have my permission to look at our code when we finish before you. Thanks Conner.

+++ src/echoserver/EchoServer.java
@@ -23,8 +23,7 @@ public class EchoServer {
 					//if (client.getInputStream() = 0) {client.close(); wait = false;}	
 						
 					// Writes the contents of the client input stream to its output stream	
-					client.getOutputStream().write(client.getInputStream().read());
-					
+					client.getOutputStream().write((byte)client.getInputStream().read());
       					}
 				}
